Prace nad złożeniem importu

+++ src/Person.java
@@ -4,8 +4,8 @@
 public class Person extends Client {
     private String firstname;
     private String surname;
-    private int idNumber;
-    public Person(String firstname, String surname, int idNumber) {
+    private long idNumber;
+    public Person(String firstname, String surname, long idNumber) {
         this.firstname = firstname;
         this.surname = surname;
         this.idNumber = idNumber;
@@ -19,7 +19,7 @@ public class Person extends Client {
         return surname;
     }
 
-    public int getIdNumber() {
+    public long getIdNumber() {
         return idNumber;
     }
 }
